:truck: rename: modules renamed
